add make_current_page_syllabus
